telegram_bot/requirements.txt
python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
gunicorn==21.2.0